+++ src/TreeSetTests.java
@@ -29,7 +29,7 @@ public class TreeSetTests {
         System.out.println("TreeSet clear time: " + clearTime);
     }
 
-    private int testAdd(int[] testCases) {
+    private long testAdd(int[] testCases) {
         long elapsedTime = 0;
         for (int num : testCases) {
             TreeSet<Integer> copySet = new TreeSet<>(set);
@@ -38,10 +38,10 @@ public class TreeSetTests {
             long end = System.nanoTime();
             elapsedTime += end - start;
         }
-        return (int) (elapsedTime / testCases.length);
+        return elapsedTime / testCases.length;
     }
 
-    private int testCheckAvailable(int[] testCases) {
+    private long testCheckAvailable(int[] testCases) {
         long elapsedTime = 0;
         for (int num : testCases) {
             long start = System.nanoTime();
@@ -49,10 +49,10 @@ public class TreeSetTests {
             long end = System.nanoTime();
             elapsedTime += end - start;
         }
-        return (int) (elapsedTime / testCases.length);
+        return elapsedTime / testCases.length;
     }
 
-    private int testRemove(int[] testCases) {
+    private long testRemove(int[] testCases) {
         long elapsedTime = 0;
         for (int num : testCases) {
             TreeSet<Integer> copySet = new TreeSet<>(set);
@@ -61,10 +61,10 @@ public class TreeSetTests {
             long end = System.nanoTime();
             elapsedTime += end - start;
         }
-        return (int) (elapsedTime / testCases.length);
+        return elapsedTime / testCases.length;
     }
 
-    private int testClear() {
+    private long testClear() {
         long elapsedTime = 0;
         for (int i = 0; i < 100; i++) {
             TreeSet<Integer> copySet = new TreeSet<>(set);
@@ -73,6 +73,6 @@ public class TreeSetTests {
             long end = System.nanoTime();
             elapsedTime += end - start;
         }
-        return (int) (elapsedTime / 100);
+        return elapsedTime / 100;
     }
 }
